Take out variants from MotionSlider component to separate file
